[android] Fix cancellation of repeating alarms, fix #1495
We noticed that we go over the limit of allowed alarms on Samsung
devices (500). This is an unlikely event to happen naturally so it must
have been a problem with cancellation of alarms. We also observed some
cases of alarms which happened when they shouldn't.

I used `adb shell dumpsys alarms` to prove it and indeed there were
cases when alarms were not removed. After debugging I noticed the
message "failed to resolve session keys" for some alarms. I narrowed it
down to the cancellation of repeating alarms. We've been using the
instance sent to us from the server which doesn't contain real fields
or keys instead of the ones saved on the disk which are complete.