[desktop] Fix downloading big missed notification response, fix #1365
Received response exceeded 130kb and JSON parsing failed because of it.
We expected that all data fits in one chunk of the "data" callback but
in this case it did not. We buffer it like we should now.

We don't handle this correctly for SSE and we should take care of that
too.