[desktop] handle mailto links on macos before app.isReady(), fix #1877
handling mailto links on macos needs special treatment since we get the
desired URL via a 'open-url' event on electron.app instead of a cli arg.
Handling them involves opening a window, which we can't do when the app
hasn't fired the 'ready' event.
The handler now sets a one-time 'ready' event listener if the app is not
ready yet.