Antiphishing system
Includes:

[desktop] Prevent "cycle of death" when web contents are crashing

In some cases when app is killed web contents have been crashing. We
tried to re-create the window and set bounds after the small delay. By
the time we tried to set bounds the window has already crashed. We
showed dialog and tried to create window again and like that untill we
crash the system.

This commit stop automatic window recreation because it's tricky to get
right and recover from invalid state