Fix race condition when loading calendar, fix #1922
This fixes race condition when loading events. We take snapshot before
loading things concurrently and then replace events map with snapshot.
This leads to overwriting one month with another. Now we replace events
only synchronously.

This also removes shortEvents from CalendarInfo because we never used
it and we did this bookkeeping for nothing. longEvents are left because
we want to load them only once.