Fix late U2F error dialog shown after login was accepted, fix #1792
We fixed this issue before for Firefox in 0edfcf3 but turns out that
Firefox and Chrome return different error codes on timeout: Firefox 1
(OTHER_ERROR), Chrome 4 (DEVICE_INELIGIBLE) and neither is really
correct as it should be 5 (TIMEOUT). This commit also silences
DEVICE_INELIGIBLE if the dialog is already hidden.