Further tidy up login error handling, fix import error
We are now just using vanilla Promise.catch, which is good because we will be removing bluebird. Also, there was a not-allowed import that was caught by the new import checker

It's also worth noting that we no longer explicitly handle ConnectionError here, because we no longer support IE, which was seemingly the only reason that we were doing so.