add override of electrons' per-url zoom factor retention
electron started to exhibit chromiums' zoom factor retention
which is saved in electron.app.getPath('userData')/Preferences
* %APPDATA%/tutanota-desktop/Preferences on win32
* ~/.config/tutanota-desktop/Preferences on linux
* ~/Library/Application Support/tutanota-desktop/Preferences on darwin

this conflicts with our own zoom handling logic.
it is saved per url and at different times apparently depending
on things like whether a window is open at the time of shutdown
or not.

this commit overrides this behaviour by maintaining the current
zoom factor in the window manager and resetting it whenever
any WebContents navigated.

also reduces the step size of scale adjustment from 10% to 5%.

fix #2917, #2317