fix contrast adjustment check, fix #3044
Originally, the contrast check was being done before we stringified the
fragment that is output by htmlSanitizer.sanitizeFragment. This was
switched around and the fragment was being consumed by
stringifyFragment before we could check it's contents.

Now we aren't calling stringify anyway, and we render the fragment
directly in the DOM.

The condition in the contrastFixNeeded check has also been slightly
simplified.

We also fixed the rendering of the error message when a mail fails
to be decrypted, which was already broken.