Handle indexer out of sync based on time since last index, fix #2453
This commit implements out of sync detection for the search index,
by tracking the time between search index updates, so that if
a user doesn't update their search index within 45 days (the TTL for
entity event updates) then we will consider the index to be out of
sync, since we may have missed updates by then.

This replaces the old method of handling out sync, which was to send
noop-events from the server.

Co-authored-by: @vitoreiji