Fix downloading files on desktop, fix #3382
Bluebird refactoring introduces a bug where files couldn't be
downloaded because DownloadItem cannot be imported from Electron. I
noticed that it is not necessary to simulate will-download event like
we did before because we don't handle it anymore (we don't use browser
download APIs for desktop anymore, I could not find when we stopped
using them).

Cancellation of saveBlob was missed before, probably because we used
plain string as an error. Now it uses CancelledError explicitly and
fileController handles it.

It is not clear to me why FileController.prototype.open() should do
saving of files instead of opening but I did not change it yet.