Refactor EntityRestCache loadRange, fix loading towards ranges, fixes #3928
It was written recursively due to having been originally written with promise chains, which made it difficult to understand/test/debug

The implementation of "loading in the direction of an existing range" was also incorrect. It would only extend in one direction.
We fixed this and added a test case