Improve how models are generated
The idea is to reduce number of separate files and imports for them to
make it easier on compiler. Unfortunately so far the gains are not as
big as we hoped but this change can be a base towards reducing direct
dependency on typeModels to that we can avoid recompiling them.
Because of that we try to avoid importing them directly and
theoretically we could reduce it even more.