Set correct owner and type for events in shared calendars, fix #3762
For the owner of a shared calendar the type of an event that was created
by  another user was wrongfully set to 'invite'.
To prevent this behavior we set the organizer of such events to be the
owner of the calendar as only the owner should be able to add guests anyways.