Fix making service requests w/o auth but w/ sessionKey, fix #4194
With offline login handling we've added a check that to get a encrypted
instance or a service response we must be logged in. It normally makes
sense but we missed the case where we pass session key manually. One
of such cases is getting giftCard information where key is encoded in
the hash part of the URL.

Now we've made an exception where we will allow making the request when
sessionKey is passed manually. This might still fail if we are not
logged in (where we want to use sessionKey to encrypt the request and
not to decrypt the response) but these cases are very rare. Better
solution would probably be to refactor the code to explicitly pass the
key for request or response (two separate fields in params) and do not
event attempt to resolve the key if that doesn't work. This would also
make CryptoFacade simpler as it would always expect to have a full
login.