Make List more safe, fix #4187
Issue #4187 looks like some kind of timing issue where we load the
list faster than we get to draw it at least once. This generally
shouldn't happen but could theoretically happen if loading is
triggered from the outside (e.g. load and select certain element).

To address this we introduced a new function updateListHeight() as this
is something that was done in a lot of places (kind of haphazardly). In
this new function and in reposition() we check if it's safe to execute
and bail out if it's not. If it's not safe yet it will be done
automatically on the first render.