fix missing connection list indicator
when offline with only a partial mail list cached, sometimes the connection lost indicator does not get rendered, even though it should. there was a race between `loadMoreIfNecessary` reading the value of `visibleElementsHeight` and `_createVirtualRows` writing to it. if it was read before it was written, then `loadMoreIfNecessary` would not try to load more from the server and would therefore not receive a connection lost error

the fix is to force `loadMoreIfNecessary` to wait until `visibleElementsHeight` has definitely been written to

pretty hacky, I'm not sure of a better way that wouldn't require larger changes

fix #4117