[desktop] Fix multiple issues with offline database deletion
fix #4344

When the database file was corrupted or got out of sync with the key
then it wasn't handled in any way, we would just fall back to the
ephemeral cache.

Now we delete the database and create it again, there's nothing else
we can do.

fix #4382

When adding another account that would override existing stored
credential we would generate a new database key but we would try to
use it with the old database.

Now instead we are forcing the creation of new database and delete
the old one directly from the worker.