Have mail dialogs made via mailto and share save as draft on close
Extend SendMailModel to allow specifying the default modified state to
allow some otherwise unmodified messages to save as a draft.

This will allow emails created via a mailto link or share (on mobile)
to default to saving as a draft even if the user made no further
modifications to the message.

Fixes #4440