Hide attachments if more than one attachment in MailViewer
Have the attachment view have a "Show all" button if there is more than
one attachment, and have this button not inline with the buttons.

Previously, the first few attachments were shown in a list with a
button to show all attachments and download all attachments, where
showing all attachments showed the remaining attachments after these
buttons. The issue with this was that the horizontal placement of these
buttons was made inconsistent, and we could result in attachments
placed on both the left and right sides of these buttons.

This change makes it so if there is more than one attachment then the
attachments are not shown until the user expands the list with the
"Show all" button. The user can then choose to download each attachment
individually or all at the same time. This makes the placement of the
"Show all" and "Save all" buttons consistent and thus easy to find.

Fixes #4479