fix unread state toggle blocking the mailviewer in one-column layout
when selecting an unread mail, we waited for the unread state being
updated before showing the mail body, which in the one-column layout
means that the view would remain on the mail list until that operation
finishes.

when you're offline, this never finishes, which means we never show the
viewer even though there might be something in the cache to display.

making the toggle operation fully async fixes this.

Co-authored-by: ivk <ivk@tutao.de>