Fix updating mail viewer when switching between emails
When we switch between mails we need to create a new viewModel because
it holds a lot of state. Originally we introduced an update only to
avoid re-creating viewModel for the same updated email. This commit
compares ids of the mail to make the decision.

Another part is fixing the flashing of the mail viewer when switching
between cached emails. The way it was done was not working as the
element we've been rendering while delaying did not actually match
any other element and it would be visible for a short amount of time.

Now we are skipping re-rendering of the mail viewer altogether during
the switching interval. This makes for a smooth, atomic transition
between two mails.

fix #4517