Fix readding external guest not restoring password field
Upon adding, removing and re-adding an external guest to a calendar
event, the password fields would not be rendered for that guest.

The "RecipientType" of the guest upon re-adding was considered
"unknown" although this should be either internal or external which is
also necessary for the "CalendarEventEditDialog" to render the password
fields.

The reason this bug occurred within the calendar but not the maileditor
is because the maileditor includes a filter which also accepts
recipients with the "unknown" type to be rendered.
However, a recipient should only have its type be "unknown" if
resolving it failed and in that case also not be rendered on screen.
We therefore can also remove the filter condition in the maileditor and
simply have the password fields rendered when the recipients resolve to
being "external".

The issue within the calendar was that the notification of the
recipient resolving is not sent within the calendar due to it also
utilizing the "SendMailModel" which only notifies the listener when
the mail itself changes, i. e. not when dealing with a calendar.
Notifying the listener of resolving the recipient regardless
fixes this issue and renders the password fields correctly everytime.

fixes #4520