Fix ExpanderPanel sizing when the child has margin-top
The issue could only be observed in some browsers (e.g. Chrome but not
FF).

When the first child of ExpanderPanel had margin-top the expander would
not fully collapse (until some redraw happens and takes the child out of
the DOM completely). This happened because collapsible margins would
cause overflow out of the child-wrapper. This kind of worked for an
overall expanded container because the wrapper itself would be move
around but it would break subtly in a collapsed state where wrapper had
height 0 yet overall panel would have the size of the margin.

We use `display: flow-root` to prevent collapsing. In addition, we added
`overflow-hidden` to the resizable part of the expander because it
should actually clip its contents and always has explicit height.