Fix invalid handling of outdated external link for saved credentials
When the new external mailbox link is generated the salt stored on the
server is updated but existing sessions are not closed. This led to a
situation where we would detect an outdated link properly for password
login but not for the stored session. As far as we can tell the check
was always missing.

As with changed password we've added a check for an outdated link. When
this happens we show the message but do not delete anything as the
session might still be valid for the up-to-date link.

fix #4184