Fix the misbehavior of the Android system back button in the settings
The system back button on Android leads from the detail search view
in one of the settings back to MailView, when it should take you back
to the settings overview. For example for user management.

Therefore the order of the if statements in WebMobileFacade.ts has been
changed to check at an earlier stage whether it is possible to
focus the previous column. This only applies if the main column is not
the focused one. So we can eventually return to the main column.
Furthermore singleColumnLayout is a prerequisite because in the
multi-column layout it is not necessary to return to the previous column
since all columns are already visible.
also fix #4621 and fix #4649

fix #4656