Catch misconfigured usage test errors, #4736
We now catch BadRequestErrors and handle
usage test stages not being registered
in case a usage test is misconfigured.

This is so misconfigured tests never affect
user experience, but it also means that we need to
test usage test implementations extra diligently
and make sure they are configured correctly.
Otherwise, we might silently run a misconfigured test
for a while and miss pings.

fixes #4736