give ViewSlider the ability to handle browser history events
if the view slider has a previous (further left) column that's not a
foreground column it can focus, it will do so and signal the window
facade that the back button event has been handled.

otherwise, if the focused column is a foreground column, it will focus
the next column to the right and signal the handled event.
this has the effect of closing the drawer menu in our application.

fix #4731

Co-authored-by: nok <nok@tutao.de>