Fix click events not being dispatched to some buttons
UpgradeSubscriptionPage would incorrectly create new component wrappers
every time. mithril wouldn't diff them correctly and would re-create
the buttons on every redraw.

This clashed with our new pointerdown handler in RootView which would
trigger the redraw. Click would be dispatched only in the next frame
when the pointer would be up. By that time the buttons would be replaced
and would never get the click event.

We replaced incorrect component creation and switched the event to
pointerup to prevent this in the future.

fix private-issues#114