Fix buggy size adjustment while resizing the month calendar vertically
When changing the height there is a slight delay for recalculating the
layout of the events. In this gap the view would overflow vertically
and show a scrollbar for a brief moment. This scrollbar would cause
a layout shift in the calendar. After the gap it would turn back.

We prevented the scrollbar with overflow-y: clip because none of the
calendar views should overflow vertically anyway.

Co-authored-by: ivk <ivk@tutao.de>