fix buggy list of notifications on ios
Previously a redraw was made after every touch or key event, setting the
correct css class. Apparently redrawing breaks the button behaviour.
A comment in a previous commit stated:
// It is important to not redraw at this point because click event may
be lost otherwise and saved login button would not be
// actually pressed. It's unclear why but preventing redraw (this way or
 setting listener manually) helps.
// It's also useless to redraw for this click handler because we just
want to add a global style definition.

Now the initial navigationType is still used, and then css classes are
removed and added to the root element on events without requiring
a redraw.

fix #4869

Co-authored-by: Willow <charlag@tutanota.com>