Fix trash, spam & subfolders getting stuck in the loading state
The bug was introduced with the async check for the subfolders of trash
and spam. To display the banner above the list we wrap both the list
and the header in ListColumnWrapper but before we would only render
the list inside the wrapper if we wanted to show the banner. With the
async folder check the value of showingSpamOrTrash changes at an
arbitrary moment of time and the list would be deleted and created at
some moment of time as well. Some timing issue of re-creating the list
would cause it to miss the state change for loading.

We worked around the issue by always rendering the wrapper which makes
the DOM structure more stable and avoids re-creating the list. A better
fix would also include changes to the List to make it more robust.

fix #5081