Use exposeLocalDelayed instead of exposeLocal for remote bridge
Remote invocations are already async but we require facade to be
there when we expose them to the other side. It doesn't cost us
anything to additionally wait for the facade itself.

The primary motivation is to make it easier to move away from
the monolithic worker and sync locators towards more async ones.