fix the unnecessary margin at the right of the news badge
The margin actually appeared for every CounterBadge, like unread
counter, news item counter, minimized draft counter. Only in the case of
the unread email counter the margin is needed. That's why I moved it
from the component and wrapped the CounterBadge in case of the news
badge in an extra div with the accurate margin.

fix #5160