improve display of short file names in attachment bubble
we can remove the right padding on the static right text in the bubble
if we use a comma.
this enables us to have no difference between the unsplit name in
the expanded bubble and the split name in the collapsed bubble
for short names that do not get ellipsed.

since the space right of the ellipsis is added by the browser, we're
unable to remove it easily.

fixes dee136a0c13c5be60188c5d2ba83c487bce4e060