try to make sure AttachmentBubble animates to height > 0
there is a possibility that we get 0 for domContent.offsetHeight
in some circumstances, but it's unclear when.
it might have something to do with the opacity: 0 hack
or because the original 24ms delay was not enough.

85 is a value that fits for a single line of attachment
name, but looks weird for longer names while keeping
the buttons accessible. if the reports continue despite this fix,
we can ask for logs.

fix #5587
fix dee136a0c13c5be60188c5d2ba83c487bce4e060